refactor: move error-warning from repository/onboarding/pr/ to repository/
