+++ lib/workers/repository/errors-warnings.ts
@@ -1,7 +1,7 @@
-import type { RenovateConfig } from '../../../../config/types';
-import { logger } from '../../../../logger';
-import type { PackageFile } from '../../../../modules/manager/types';
-import { emojify } from '../../../../util/emoji';
+import type { RenovateConfig } from '../../config/types';
+import { logger } from '../../logger';
+import type { PackageFile } from '../../modules/manager/types';
+import { emojify } from '../../util/emoji';
 
 export function getWarnings(config: RenovateConfig): string {
   if (!config.warnings.length) {
