+++ tsconfig.strict.json
@@ -48,7 +48,7 @@
     "lib/workers/repository/onboarding/branch/rebase.ts",
     "lib/workers/repository/onboarding/pr/base-branch.ts",
     "lib/workers/repository/onboarding/pr/config-description.ts",
-    "lib/workers/repository/onboarding/pr/errors-warnings.ts",
+    "lib/workers/repository/errors-warnings.ts",
     "lib/workers/repository/onboarding/pr/index.ts",
     "lib/workers/repository/onboarding/pr/pr-list.ts",
     "lib/workers/repository/process/deprecated.ts",
